Move ColumnInfoInput
